remove OmniBox.tsx
